pip requirements.txt
pip requirements:
python-telegram-bot
sklearn
nltk